Use adjusted total allele counts instead of unadjusted
The default AC and AN are the "unadjusted" allele counts. This
contrasts with the default sub-population AC and AN (i.e. AC_AFR,
AN_AFR, etc.). which use adjusted allele counts. Therefore this
script has been modified to use the adjustd allele counts.

Also renamed the .includes/ dir to includes/